requirements.txt
docker_py==1.10.6
Flask==3.0.2
Flask_Cors==4.0.0
gdown==5.1.0
keras==2.11.0
kneed==0.8.5
kserve==0.12.0
matplotlib==3.7.1
numpy==1.24.2
nvidia_ml_py==12.535.133
opencv_python==4.7.0.72
Pillow==10.3.0
pynvml==11.5.0
python-dotenv==1.0.1
requests==2.25.1
scikit_learn==1.4.1.post1
seaborn==0.13.2
tensorflow==2.11.0
transformers==4.38.1
